requirements.txt
python-docx
openpyxl
ttkbootstrap
pyinstaller
